pyproject.toml
[tool.poetry]
name = "spacetagram"
version = "0.1.0"
description = ""
authors = ["nvalle0624 <nvalle0624@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.9"
Django = "^3.2.7"
djangorestframework = "^3.12.4"

[tool.poetry.dev-dependencies]
autopep8 = "^1.5.7"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
